# Vectorized hash grouping