… Cargo.toml
… benchmarks/src/bin/tpch.rs
… datafusion-cli/Cargo.lock
… datafusion/core/Cargo.toml
… datafusion/core/src/datasource/physical_plan/parquet.rs
… datafusion/core/src/physical_optimizer/repartition.rs
… datafusion/core/src/physical_plan/aggregates/bounded_aggregate_stream.rs
… datafusion/core/src/physical_plan/aggregates/mod.rs
… datafusion/core/src/physical_plan/aggregates/row_hash.rs
… datafusion/core/src/physical_plan/aggregates/row_hash2.rs
… datafusion/core/src/physical_plan/joins/nested_loop_join.rs
… datafusion/core/tests/data/empty_files/some_empty/b.csv
… datafusion/core/tests/data/empty_files/some_empty/d.csv
… datafusion/core/tests/data/one_col.csv
… datafusion/core/tests/data/wide_rows.csv
… datafusion/core/tests/fuzz_cases/aggregate_fuzz.rs
… datafusion/core/tests/sqllogictests/src/engines/datafusion/error.rs
… datafusion/core/tests/sqllogictests/test_files/aggregate.slt
… datafusion/core/tests/sqllogictests/test_files/insert.slt
… datafusion/core/tests/sqllogictests/test_files/scalar.slt
… datafusion/core/tests/sqllogictests/test_files/tpch/q1.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q10.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q11.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q12.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q13.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q14.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q15.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q16.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q17.slt.part
… datafusion/core/tests/sqllogictests/test_files/tpch/q18.slt.part