(draft until I rerun the benchmark numbers)

# Which issue does this PR close?

Part of https://github.com/apache/arrow-datafusion/issues/6889
Closes https://github.com/apache/arrow-datafusion/pull/6800 
Closes https://github.com/apache/arrow-datafusion/issues/4973
Closes https://github.com/apache/arrow-datafusion/issues/956
Closes https://github.com/apache/arrow-datafusion/issues/846
Closes https://github.com/apache/arrow-datafusion/issues/418


# Rationale for this change

Much faster grouping performance and lower memory usage for large numbers of groups

## TPCH (SF1) 
```
--------------------
Benchmark tpch_mem.json
--------------------
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃ main_base ┃ alamb_fast_gby_hash ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │  521.08ms │            478.59ms │ +1.09x faster │
│ QQuery 2     │  238.03ms │            148.34ms │ +1.60x faster │
│ QQuery 3     │  171.12ms │            156.71ms │ +1.09x faster │
│ QQuery 4     │  113.90ms │            111.03ms │     no change │
│ QQuery 5     │  366.87ms │            359.57ms │     no change │
│ QQuery 6     │   39.49ms │             37.80ms │     no change │
│ QQuery 7     │  858.86ms │            854.47ms │     no change │
│ QQuery 8     │  234.09ms │            219.26ms │ +1.07x faster │
│ QQuery 9     │  564.17ms │            535.61ms │ +1.05x faster │
│ QQuery 10    │  323.68ms │            323.61ms │     no change │
│ QQuery 11    │  171.33ms │            153.55ms │ +1.12x faster │
│ QQuery 12    │  154.57ms │            156.19ms │     no change │
│ QQuery 13    │  562.33ms │            286.53ms │ +1.96x faster │
│ QQuery 14    │   45.14ms │             47.25ms │     no change │
│ QQuery 15    │   81.27ms │             49.87ms │ +1.63x faster │
│ QQuery 16    │  187.83ms │            154.13ms │ +1.22x faster │
│ QQuery 17    │ 2597.93ms │            893.75ms │ +2.91x faster │
│ QQuery 18    │ 2239.90ms │           1505.12ms │ +1.49x faster │
│ QQuery 19    │  156.73ms │            145.82ms │ +1.07x faster │
│ QQuery 20    │  779.21ms │            309.94ms │ +2.51x faster │
│ QQuery 21    │ 1049.99ms │            997.71ms │     no change │
│ QQuery 22    │   85.38ms │             85.93ms │     no change │
└──────────────┴───────────┴─────────────────────┴───────────────┘
```



## ClickBench results

```
+-------+----------+-------------------+--------------+----------+-------------------+--------------+
| query | main_min | fast_gby_hash_min | cmp          | main_avg | fast_gby_hash_avg | cmp21        |
+-------+----------+-------------------+--------------+----------+-------------------+--------------+
| 1     | 0.439    | 0.41              | NO CHANGE    | 1.24     | 1.23              | NO CHANGE    |
| 2     | 0.199    | 0.2               | NO CHANGE    | 0.24     | 0.23              | NO CHANGE    |
| 3     | 0.38     | 0.46              | NO CHANGE    | 0.42     | 0.7               | NO CHANGE    |
| 4     | 0.386    | 0.4               | NO CHANGE    | 0.63     | 0.65              | NO CHANGE    |
| 5     | 8.945    | 3.87              | 2.31x faster | 9.06     | 4.12              | 2.2x faster  |
| 6     | 5.782    | 4.0               | NO CHANGE    | 7.59     | 4.11              | 1.85x faster |
| 7     | 0.375    | 0.19              | 1.94x faster | 0.41     | 0.22              | 1.86x faster |
| 8     | 0.401    | 0.21              | 1.92x faster | 0.46     | 0.25              | 1.88x faster |
| 9     | 10.36    | 4.58              | 2.26x faster | 16.85    | 4.75              | 3.55x faster |
| 10    | 8.818    | 8.3               | NO CHANGE    | 9.07     | 8.46              | NO CHANGE    |
| 11    | 1.194    | 1.06              | NO CHANGE    | 1.3      | 1.17              | NO CHANGE    |
| 12    | 1.286    | 1.13              | NO CHANGE    | 1.38     | 1.24              | NO CHANGE    |
| 13    | 5.252    | 3.44              | 1.53x faster | 5.37     | 3.53              | 1.52x faster |
| 14    | 10.012   | 6.52              | 1.54x faster | 10.23    | 6.8               | 1.5x faster  |
| 15    | 5.555    | 3.64              | 1.53x faster | 5.72     | 3.77              | 1.52x faster |
| 16    | 10.95    | 4.19              | 2.61x faster | 11.15    | 4.38              | 2.54x faster |
| 17    | 15.721   | 8.37              | 1.88x faster | 16.27    | 8.72              | 1.87x faster |
| 18    | 14.158   | 8.16              | 1.73x faster | 14.38    | 8.24              | 1.75x faster |
| 19    | 33.824   | 16.62             | 2.04x faster | 37.64    | 17.31             | 2.17x faster |
| 20    | 0.29     | 0.3               | NO CHANGE    | 0.52     | 0.51              | NO CHANGE    |
| 21    | 3.696    | 3.69              | NO CHANGE    | 5.2      | 5.17              | NO CHANGE    |
| 22    | 4.591    | 4.58              | NO CHANGE    | 6.2      | 6.22              | NO CHANGE    |
| 23    | 13.191   | 13.24             | NO CHANGE    | 15.07    | 15.22             | NO CHANGE    |
| 24    | 31.688   | 32.11             | NO CHANGE    | 36.17    | 37.21             | NO CHANGE    |
| 25    | 1.824    | 1.85              | NO CHANGE    | 2.23     | 2.24              | NO CHANGE    |
| 26    | 1.499    | 1.49              | NO CHANGE    | 1.59     | 1.58              | NO CHANGE    |
| 27    | 1.883    | 1.85              | NO CHANGE    | 2.57     | 2.26              | NO CHANGE    |
| 28    | 6.902    | 5.39              | NO CHANGE    | 9.19     | 6.61              | NO CHANGE    |
| 29    | 15.324   | 15.13             | NO CHANGE    | 19.39    | 15.63             | NO CHANGE    |
| 30    | 1.672    | 1.62              | NO CHANGE    | 1.71     | 1.69              | NO CHANGE    |
| 31    | 6.689    | 3.13              | 2.14x faster | 7.02     | 4.23              | 1.66x faster |
| 32    | 9.18     | 8.13              | NO CHANGE    | 9.73     | 9.17              | NO CHANGE    |
| 33    |          |                   |              |          |                   |              |
| 34    | 22.425   | 18.66             | NO CHANGE    | 23.43    | 29.08             | NO CHANGE    |
| 35    | 23.101   | 18.83             | NO CHANGE    | 23.69    | 19.46             | NO CHANGE    |
| 36    | 10.674   | 5.15              | 2.07x faster | 10.85    | 5.26              | 2.06x faster |
| 37    | 0.845    | 0.79              | NO CHANGE    | 0.93     | 0.84              | NO CHANGE    |
| 38    | 0.465    | 0.43              | NO CHANGE    | 0.5      | 0.47              | NO CHANGE    |
| 39    | 0.386    | 0.37              | NO CHANGE    | 0.45     | 0.42              | NO CHANGE    |
| 40    | 1.546    | 1.59              | NO CHANGE    | 1.66     | 1.68              | NO CHANGE    |
| 41    | 0.222    | 0.2               | NO CHANGE    | 0.27     | 0.24              | NO CHANGE    |
| 42    | 0.207    | 0.19              | NO CHANGE    | 0.24     | 0.22              | NO CHANGE    |
| 43    | 0.237    | 0.22              | NO CHANGE    | 0.27     | 0.24              | NO CHANGE    |
+-------+----------+-------------------+--------------+----------+-------------------+--------------+
```

<details><summary>Methodology</summary>
<p>
1. Run `bash.run.sh` this script: https://github.com/ClickHouse/ClickBench/tree/main/datafusion
2. Save the resulting `result.csv`:  [result-fast_gby_hash.csv](https://github.com/apache/arrow-datafusion/files/12014914/result-fast_gby_hash.csv) [result-main.csv](https://github.com/apache/arrow-datafusion/files/12014915/result-main.csv)
3. Run report: `datafusion-cli -f report.sql`

report.sql
```sql
create external table main(query int, run int, time float) stored as csv location 'result-main.csv';
create external table fast_gby_hash(query int, run int, time float) stored as csv location 'result-fast_gby_hash.csv';

--Comparison

SELECT
  main.query as query,
  main.min_time as main_min,
  round(fast_gby_hash.min_time, 2) as fast_gby_hash_min,

  CASE
    -- ignore results that are within 5%
    WHEN abs(1 - abs(main.min_time / fast_gby_hash.min_time)) < 0.5
      THEN 'NO CHANGE'
    WHEN main.min_time > fast_gby_hash.min_time
      THEN round(main.min_time / fast_gby_hash.min_time, 2) || 'x faster'
    ELSE
      round(fast_gby_hash.min_time / main.min_time, 2) || 'x slower'
  END as cmp,

  round(main.avg_time, 2) as main_avg,
  round(fast_gby_hash.avg_time,2) as fast_gby_hash_avg,
  CASE
    -- ignore results that are within 5%
    WHEN abs(1 - abs(main.avg_time / fast_gby_hash.avg_time)) < 0.5
      THEN 'NO CHANGE'
    WHEN main.avg_time > fast_gby_hash.avg_time
      THEN round(main.avg_time / fast_gby_hash.avg_time, 2) || 'x faster'
    ELSE
      round(fast_gby_hash.avg_time / main.avg_time, 2) || 'x slower'
  END as cmp21
FROM
  (select query, min(time) as min_time, avg(time) as avg_time from main group by 1) as main
    JOIN
  (select query, min(time) as min_time, avg(time) as avg_time from fast_gby_hash group by 1) as fast_gby_hash
    USING (query)
ORDER BY
  query
;

```


</p>
</details> 

# What changes are included in this PR?
The code in this PR was written by myself and @Dandandan 

- [x] Rewrite `GroupedHashAggregateStream` to use vectorized / multi-group updates
- [x] A new `GroupsAccumulator` trait with the new vectorized API for managing and updating group state
- [x] An generic implementation of `GroupsAccumulator` for all aggregators that have `RowAccumulator` variants
- [x] Fuzz testing of new code `accumulate` function
- [x]  An adapter that implements `GroupsAccumulator` in terms of `Accumulator` (for slower, but simpler accumulators)
- [x] Significant documentation for all new code

Here is the list of `RowAccumulator`s (aka accumulators that have specialized implementations). 

- [x] `CountRowAccumulator`
- [x] `MaxRowAccumulator`
- [x] `MinRowAccumulator`
- [x] `AvgRowAccumulator`
- [x] `SumRowAccumulator`
- [x] `BitAndRowAccumulator`
- [x] `BitOrRowAccumulator`
- [x] `BitXorRowAccumulator`
- [x] `BoolAndRowAccumulator`
- [x] `BoolOrRowAccumulator`

# Follow on work
I plan to remove the GroupedAggregateStream (and all RowAccumulators) as a follow on PR

# Are these changes tested?
Yes -- both new and existing tests 

# Are there any user-facing changes?
Much faster performance -- see above