@alamb 2023-07-11
>  In the long run, will RowAccumulators still be used, or did they become obsolete with this PR. It seems to me that they are not used anymore, maybe I am missing something.

This is an excellent question @mustafasrepo 

My plan is to remove `RowAccumulators` (ideally later this week) -- however, they are still still used by `BoundedAggregateStream`, so I need to complete https://github.com/apache/arrow-datafusion/issues/6798 first. I have a plan for this, I just need to write the code for it now. 

I listed out my plan on the description https://github.com/apache/arrow-datafusion/issues/6889 -- but I realize that probably got list in the flurry of activity and tickets.