@alamb 2023-07-12
> Once this merges and the old code is removed,

FWIW I am actively coding on https://github.com/apache/arrow-datafusion/issues/6798 right now