@alamb 2023-07-13
While writing tests (specifically https://github.com/apache/arrow-datafusion/pull/6939) I found an error related to a discrepancy in what min/max accumulator types were supported. 

I also found a bug in how these new aggregates are  propagating nulls. I am working on a fix

```
❯ create table foo (x int) as values (1), (2), (3), (null);;
0 rows in set. Query took 0.027 seconds.
❯ select * from foo;
+---+
| x |
+---+
| 1 |
| 2 |
| 3 |
|   |
+---+
4 rows in set. Query took 0.005 seconds.
❯ select min(x) from foo;
+------------+
| MIN(foo.x) |
+------------+
| 1          |
+------------+
```

I will fix this as well as update our test coverage