@alamb 2023-07-13
I wrote  some additional tests and actually found another bug (related to bitand): https://github.com/apache/arrow-datafusion/pull/6952