@alamb 2023-07-13
https://github.com/apache/arrow-datafusion/pull/6904/commits/a98b6a0a25bd684fa16d6b314c550bb00c943e2b and https://github.com/apache/arrow-datafusion/pull/6904/commits/5ab75b1be7617771125ae9aec5095297bc64100d are new -- the rest of the commits are from merging. 

I may just merge this PR in as is and then fix the remaining bugs on master to avoid keeping it open for too much longer