@alamb 2023-07-13
My most recent changes have slowed things down a little, but I have an idea of how to get the performance back. Let's merge this and proceed with development on main.

Thank you everyone for your help and encouragement

```
--------------------
Benchmark tpch_mem.json
--------------------
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃ alamb_fast_gby_hash_pre ┃ alamb_fast_gby_hash ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │                484.92ms │            536.15ms │  1.11x slower │
│ QQuery 2     │                148.89ms │            153.45ms │     no change │
│ QQuery 3     │                165.97ms │            162.14ms │     no change │
│ QQuery 4     │                114.07ms │            127.66ms │  1.12x slower │
│ QQuery 5     │                366.55ms │            378.28ms │     no change │
│ QQuery 6     │                 41.96ms │             38.40ms │ +1.09x faster │
│ QQuery 7     │                868.25ms │            895.61ms │     no change │
│ QQuery 8     │                246.07ms │            241.03ms │     no change │
│ QQuery 9     │                572.43ms │            563.44ms │     no change │
│ QQuery 10    │                317.69ms │            311.76ms │     no change │
│ QQuery 11    │                160.81ms │            167.52ms │     no change │
│ QQuery 12    │                156.69ms │            174.96ms │  1.12x slower │
│ QQuery 13    │                284.68ms │            312.62ms │  1.10x slower │
│ QQuery 14    │                 48.32ms │             50.55ms │     no change │
│ QQuery 15    │                 50.41ms │             52.45ms │     no change │
│ QQuery 16    │                149.03ms │            148.36ms │     no change │
│ QQuery 17    │                812.65ms │            882.47ms │  1.09x slower │
│ QQuery 18    │               1560.64ms │           1575.56ms │     no change │
│ QQuery 19    │                154.62ms │            162.10ms │     no change │
│ QQuery 20    │                305.76ms │            305.81ms │     no change │
│ QQuery 21    │               1028.46ms │           1057.46ms │     no change │
│ QQuery 22    │                 92.21ms │             88.00ms │     no change │
└──────────────┴─────────────────────────┴─────────────────────┴───────────────┘
```