@mingmwang 2023-07-13
@alamb @Dandandan @ozankabak 
After this PR, the bottleneck of tpch q17 and q18 is the join ordering.  I will start work on the CBO stats estimation from next week and enhance the join ordering(Join Selection rule).