@alamb 2023-07-13
> After this PR, the bottleneck of tpch q17 and q18 is the join ordering. I will start work on the CBO stats estimation from next week and enhance the join ordering(Join Selection rule).

Awesome -- thank you @mingmwang 

cc @ozankabak who I think has been thinking about join ordering as well